adding --dry-run option
